// frontend/src/app/models/user.ts
export class User {
    id?: number;
    name?: string;
    role?: string;
    avatar?: string;
    active?: number;
    email?: string;
    password?: string;
    code?: string;
    direccion?: string;
    provincia?: string;
    municipio?: string;
    codigo_postal?: string;
    direccion_alt?: string;
    provincia_alt?: string;
    municipio_alt?: string;
    codigo_postal_alt?: string;
    telefono?: string;
    pais?: string;
    pais_alt?: string;
    
    constructor(

    ) { };
}